moved cpu temp to CpuTempComponent class
