--- iGlance/iGlance/Components/CPU.swift
@@ -1,9 +1,0 @@
-//
-//  CPU.swift
-//  iGlance
-//
-//  Created by Dominik on 27.10.18.
-//  Copyright © 2018 iGlance Corp. All rights reserved.
-//
-
-import Foundation
